feat : 유저 로그인 필터 추가
설명 : 로그인 되지 않은 유저는 특정 사이트 접근 시 로그인 페이지로 리디렉트한다.
적용 범위 : 글 삭제, 글 상세 조회, 글 목록 조회

+++ src/main/webapp/WEB-INF/board/GetBoard.jsp
@@ -3,6 +3,7 @@
 
 <%
   BoardVO board = (BoardVO) request.getAttribute("board");
+  UserVO userVO = (UserVO) request.getSession().getAttribute("user");
 %>
 
 <%@ include file="../layout/header.jsp" %>
@@ -40,9 +41,9 @@
     </table>
   </form>
   <hr>
-<%--  <% if(userVO.getRole().equals("ADMIN")) {%>
+<% if(userVO.getRole().equals("ADMIN")) {%>
   <a href='deleteBoard.do?id=<%=board.getId()%>'>글 삭제</a>
-  <% } %> --%>
+  <% } %>
 
 </center>
